requirements.txt
setuptools==38.4.0
numpy==1.13.3
scikit-learn==0.19.1
scipy==1.0.0
tensorflow==1.2.1
mysql-connector-python>=2.1.0